refactor(portal): extract job artifact catalog helpers
